fix(lab): rref boundary strict-> + fuzz catch_unwind + conformal alpha-tolerance [br-1j14du br-ipejce br-9u4ext]
Three coordinated lab-correctness fixes flagged by the prior /multi-pass
audit on src/lab/*:

br-asupersync-1j14du — RRefAccessOracle PostCloseAccess at-boundary

  on_rref_access used time >= close_time, so a virtual-time access
  stamped exactly at close_time produced a spurious PostCloseAccess
  violation. Production resolves at-boundary access via the runtime's
  lock; the oracle should not fire on tick equality alone. Switched
  to strict time > close_time. Callers that genuinely race with close
  should keep using on_rref_access_with_witness, which gates on
  observed concurrency rather than tick arithmetic.

  Tests: at-boundary access produces 0 violations;
  strictly-after-close still fires.

br-asupersync-ipejce — FuzzHarness::run_single panic isolation

  The fuzz harness invoked the user-supplied test closure directly:
  the first panic — the most interesting outcome of any fuzz
  campaign — aborted the entire campaign and attributed the crash
  to the harness rather than the asupersync invariant violation it
  represented. Wrapped the test() call in
  std::panic::catch_unwind(AssertUnwindSafe(...)) and converted
  panic payloads (downcast for &'static str / String) into a new
  InvariantViolation::TestPanic { message } variant. Updated
  violation_category in fuzz.rs and dual_run.rs to surface
  test_panic in violation_counts. Added Display branch.

  Tests: a panicking closure produces a TestPanic finding whose
  message contains the panic payload, and run_single returns
  normally instead of aborting.

br-asupersync-9u4ext — CalibrationReport tolerance derivation

  is_well_calibrated and miscalibrated_invariants used a fixed
  5-percentage-point absolute slack — roughly 5x the default alpha
  itself, wide enough that a system whose anomaly rate had climbed
  to 9% was still reported well-calibrated. The conformal-prediction
  guarantee operators believe they are getting was effectively
  defeated.

  Replaced with calibration_tolerance() = max(alpha/5, EPSILON).
  At alpha=0.05 this is 0.01 (1pp), restoring the guarantee. The
  tolerance is exposed as a public method so harness reports can
  show the same number that drives the pass/fail decision.

  Tests: alpha-derived tolerance values; 90% coverage at
  alpha=0.05 now correctly flags miscalibrated; 94% sits exactly
  on the new boundary; 95% comfortably passes.

+++ src/lab/conformal.rs
@@ -1662,13 +1662,13 @@ mod tests {
     // alpha/5 (1pp at the default alpha=0.05).
     // ===================================================================
 
-    fn report_with(alpha: f64, total: u64, covered: u64) -> CalibrationReport {
+    fn report_with(alpha: f64, total: usize, covered: usize) -> CalibrationReport {
         CalibrationReport {
             prediction_sets: Vec::new(),
             coverage: BTreeMap::new(),
             overall_coverage: CoverageTracker { total, covered },
             alpha,
-            calibration_samples: total as usize,
+            calibration_samples: total,
         }
     }
